// DrTech.Web/src/app/pages/request/dto/reuse-dto.ts
export class ReuseDTO {
  id: string;
  longitude: number;
  latitude: number;
  fileName: string;
  Type: number;
  description: string;
  userId: string;
  donateTo: number;
  deliveryDate: string;
  status: number;
  city: number;
  pinType: string;
  file: any;
  greenPoints:number;
  updatedDate:Date

}